requirements.txt
fastapi
uvicorn
httpx
python-dotenv
yt-dlp
pydantic
numpy
faiss-cpu

